+++ src/main/webapp/constants.js
@@ -1,0 +1,17 @@
+const colorScheme = {
+  "unmodifiedNodeColor": "blue",
+  "addedObjectColor": "green",
+  "deletedObjectColor": "red",
+  "modifiedNodeColor": "yellow",
+  "unmodifiedEdgeColor": "grey",
+  "labelColor": "white"
+};
+// Sets the opacity constants for different types of objects in the graph
+const opacityScheme = {
+  "deletedObjectOpacity": 0.25
+};
+
+module.exports = {
+  colorScheme, 
+  opacityScheme
+}
